--- app/application/forms/init.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-Dir.glob("#{__dir__}/*.rb").each do |file|
-  require file
-end
