+++ app/application/init.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-folders = %w[forms services controllers]
+folders = %w[controllers requests services]
 folders.each do |folder|
   require_relative "#{folder}/init"
 end
