+++ app/presentation/init.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-folders = %w[view_objects]
+folders = %w[representers responses]
 folders.each do |folder|
   require_relative "#{folder}/init.rb"
 end
