--- app/presentation/views_html/details.slim
@@ -1,52 +1,0 @@
-.container
-  .row
-    / h1 #{query}
-    / p #{jobs[0].description}
-    h2.my-3
-      span.fst-italic id="title" =< "#{skillset.query}"
-    a.link-secondary id="to_last_page" href="/?query=#{skillset.query}"
-      i.fas.fa-share.mx-1
-      | search for other jobs
-      
-  .row
-    .col-4
-      / Skills Required card
-      .card
-        .card-header Skills Required
-        .card-body
-          ul.list-group.list-group-flush id="skill_list"
-            - skillset.skills.each do |skill|
-              li.list-group-item
-                span id="skill_name"
-                  = skill.name
-                span.badge.mx-2.bg-light.rounded.text-black.float-end id="skill_count"
-                  = skill.count
-    .col-8
-      / Salary card
-      .card
-        .card-header Salary
-        .card-body
-          ul.list-group.list-group-flush id="salary_list"
-            li.list-group-item
-              | Maximum Year Salary:
-              span< = skillset.max_salary
-            li.list-group-item
-              | Minimum Year Salary:
-              span< = skillset.min_salary
-  
-  .row
-    / Job Vacancy cards
-    div id="vacancies"
-      - skillset.jobs.each do |job|
-        .card.my-5.p-3
-          p.card-subtitle.text-muted
-            i.fas.fa-map-marker-alt.mx-1
-            span = job.location
-          h4.px-2.card-title = job.title
-          hr.my-2
-          p.p-3.px-4.card-text = job.brief
-        
-
-  / Navigator
-  a.position-fixed.bottom-0.end-0.m-3.fs-2.link-secondary href='#'
-    i.fas.fa-chevron-circle-up
