--- app/presentation/views_html/flash_bar.slim
@@ -1,9 +1,0 @@
-- if flash[:error].nil?
-- else
-  div class="alert alert-danger" id="flash_bar_danger" = flash[:error]
-  end
-
-- if flash[:notice].nil?
-- else
-  div class="alert alert-success" id="flash_bar_success" = flash[:notice]
-  end
