--- app/presentation/views_html/index.slim
@@ -1,16 +1,0 @@
-div.row 
-  .row.text-center
-    h1.display-1 S-killer
-    p Know your job. Know your skills.
-
-  .row
-    form action='/details' method='POST'
-      .card.p-3.border-none.bg-light
-        label.form-label.card-title.fs-5 for='query' 
-          | Enter job title or skills to analyze:
-        .input-group.mb-3
-          input.form-control type='text' name='query' placeholder='data scientist' value="#{query}"
-          button.btn.btn-primary type='submit'
-            | submit
-
-include :flash_bar
