--- app/presentation/views_html/layout.slim
@@ -1,35 +1,0 @@
-doctype html
-html lang='en'
-  head 
-    meta charset='utf-8'
-    meta name='viewport' content="width=device-width, initial-scale=1, shrink-to-fit=no"
-
-    title S-killer 
-    
-    / Bootstrap
-    link{
-      href='https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css'
-      rel='stylesheet' integrity='sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3' 
-      crossorigin='anonymous'
-    }
-    / Font Awesome
-    link{
-      href='https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.14.0/css/all.min.css'
-      rel='stylesheet' type='text/css' crossorigin='anonymous'
-    }
-
-  body
-    nav.navbar.navbar-dark.bg-primary
-      div.px-3
-        a.navbar-brand href='/' S-killer
-
-    .container.p-3.py-5
-      == yield
-
-    .container.py-3
-      div.my-4.border-bottom
-      div.text-center.fst-italic
-        p.m-1.text-muted 2021 KeroKnows, NTHU
-        a.m-1.link-secondary href='https://github.com/KeroKnows/S-killer'
-          i.fab.fa-github.mx-1
-          | GitHub
